project_app: documentation
